CheckboxGroup component logic update
