+++ public/controllers/register-agent/components/os-card.tsx
@@ -1,4 +1,5 @@
-import React from 'react';
+//
+import React, { useState } from 'react';
 import {
   EuiCard,
   EuiFlexGroup,
@@ -7,26 +8,41 @@ import {
   EuiSpacer,
 } from '@elastic/eui';
 import { REGISTER_AGENT_DATA } from '../utils/register-agent-data';
-import { SwitchComponent } from './switch';
+import { CheckboxGroupComponent } from './checkbox-group';
 
-export const OsCard = () => (
-  <div>
-    <EuiSpacer size='s' />
-    <EuiFlexGroup gutterSize='l' wrap>
-      {REGISTER_AGENT_DATA.map((data, index) => (
-        <EuiFlexItem key={index}>
-          <EuiCard
-            icon={<EuiIcon size='xl' type='logoLogging' />}
-            title='Bordered'
-            display='plain'
-            hasBorder
-            description='This one has a plain background color and border.'
-            onClick={() => {}}
-          >
-            <SwitchComponent data={data} cardIndex={index} />
-          </EuiCard>
-        </EuiFlexItem>
-      ))}
-    </EuiFlexGroup>
-  </div>
-);
+export const OsCard = () => {
+  const [selectedOption, setSelectedOption] = useState<string | undefined>(
+    undefined,
+  );
+
+  const handleOptionChange = (optionId: string) => {
+    setSelectedOption(optionId);
+  };
+
+  return (
+    <div>
+      <EuiSpacer size='s' />
+      <EuiFlexGroup gutterSize='l' wrap>
+        {REGISTER_AGENT_DATA.map((data, index) => (
+          <EuiFlexItem key={index}>
+            <EuiCard
+              icon={<EuiIcon size='xl' type='logoLogging' />}
+              title='Bordered'
+              display='plain'
+              hasBorder
+              description='This one has a plain background color and border.'
+              onClick={() => {}}
+            >
+              <CheckboxGroupComponent
+                data={data.architecture}
+                cardIndex={index}
+                selectedOption={selectedOption}
+                onOptionChange={handleOptionChange}
+              />
+            </EuiCard>
+          </EuiFlexItem>
+        ))}
+      </EuiFlexGroup>
+    </div>
+  );
+};
